return error on failed client metadata lookup
When a client metadata call failed becuase of an incorrect but not blank
client_id or mismatched client_id/client_secret pair, we were returning
a null, which we would then try to dereference.   Now, handle the error
gracefully by returning a Bad Request error.  I've opted to use the same
message for both the wrong client id and mismatched credentials errors,
so that we don't leak information on which client id are legit.