Handle remote principal authenctication errors better
`notifyRemotePrincipalAuthenticated()` has a `try...catch...` that handles
all the exceptions in one place. This makes logging and errors much less
helpful/useful than it could/should be. In addition, it doesn't catch the
error where "names" are not provided.

This commit (1) splits the `try...catch...` to each independent functional
parts and uses more specific error messages, (2) catches "missing names"
errors instead of delegating this to OSF, (3) use `opt` instead of `get`
to obtain JSON objects and strings to avoid `JSONException` and extra `if`
statement.