Update source and target version from 1.7 to 1.8
Java SE 8 is strongly compatible with previous versions of the Java
platform. Almost all existing programs should run on Java SE 8 with-
out modification. However, there are some minor potential incompati-
bilities in the JRE and JDK that involve rare circumstances that are
documented in the reference below for completeness.

Ref: https://www.oracle.com/technetwork/java/javase/8-compatibility-guide-2156366.html