Reinstate "REMOTE_USER" as a auth ID requirement
The header "REMOTE_USER" is still REQUIRED as an id for institution
users although 1) OSF and CAS already rely on `username` for ID pur-
pose and 2) `notifyRemotePrincipalAuthenticated()` guarantees that
`username` is provided. The reason for keeping this extra check in
place is that `username` is not always the identifier since it can
be `eppn`, `mail` or other attributes. "REMOTE_USER" is defined as
`REMOTE_USER="eppn uid persistent-id targeted-id"`, in which every
attribute can be considered as the users' institution identity. In
short, CAS requires at least one of the four ID attributes `eppn`,
`uid`, `persistent-id`, `targeted-id` in addition to attributes that
are mapped to `username` and `fullname`. Please see our Shibboleth
server's configuration (private) for detailed mapping for attributes.
The following link provides the best practice of using "REMOTE_USER":
wiki.shibboleth.net/confluence/display/SHIB2/NativeSPAttributeAccess