Use tokenPk and scopePk to name the primary key id
For access token, the model has the integer id (PK) and a string
token_id. Unlike python, the name tokenId in Java is ambiguous.
Given that we have used tokenId for token_id, use tokenPk for id.

For scope, the model has the integer id (PK) and a _id (GUID).
Similarly, the name scopeId is ambiguous. Use scopePk instead.