Rewrite comments for OAuth constants
The naming of OAuth constants and their JavaDoc comments were far
from accurate and can be quite confusing at times. Although sadly
we have to live with the naming, the doc was rewritten so that it
accurately, concisely and consistly describes what each constant
is and how/where they are used.