Throw client exceptions in client action instead of auth handlers
The initial design was to throw client specific exceptions in the
client auth handler. However, we discovered that exceptions ended
up being caught and re-thrown as a generic CAS auth exception. To
fix this issue, the throwing client exceptions job has been moved
to the client action, which is the last step before the failures
are handled by the primary auth exception handler.

In addtion, created a new helper class that selects the correct
exception class to throw / return based on the simple class name
of the client. The class name is used instead of the client name
since each CAS client has its own unique client name.