Prefix COS modification copyright to maven-overlayed classes
OSF CAS uses maven overlay to customize / override many classes
from external packages. These classes already have an Apache 2.0
license header from their original creator so we simply added the
COS modification copyright on top of it. The copyright year
indicates the time of first overlay. In addition, we fixed the
"since" tag in the JavaDoc for the classes. This tag should be
the same as the original one indicating the time of creation.