Ignore multi-value attributes gracefully for cas-pac4j inst SSO
Multi-value attributes broke institution SSO when they are added
to the delegation attributes map. The map expects String as the
value but multi-value ones are of type ArrayList. Type cast error
is thrown in this case. This commit ignores such attributes with
error level logs to inform sentry. As long as the uid, mail and
names attributes are single-value, SSO can proceed w/o a problem.

In addition, logging are improved to include more information such
as the clientName (which institution) and principalID (which user).