Fixe key uniqueness for institution login URL map
With the introduction of shared-SSO between institutions, the login
URL (being the key of the map) is no longer unique. To fix this, a
fragment is added to each login URL with the institution ID in order
to preserve the uniqueness. For example, both Brown Univ. and The
Policy Lab at Brown use the same login endpoint but different URL.

* Brown -  https://brown.login.url#brown
* The Policy Lab - https://brown.login.url#thepolicylab

The fragment is removed by the front-end script later on the page.