Add institution SSO attribute mapping for kuleuven
Note one: Default attributes mapping are used since KU Leuven uses
the Belnet federation, which is the Belgium participant of eduGain.
As expected, their IdP metadata is found in the metadata aggregate
of InCommon, which is the U.S. participant of eduGain.

Note two: as usual, this change only affects local development but
it serves as the reference for private server settings update.